Xcode project monkeying
I had been getting some unexplained crashes on x86_64.  Some digging
revealed it had to do with vector extensions.  If I built vpxenc with
the project set to AVX 2 and added compiler flags to build the one AVX
512 file vpxenc worked, but I still got crashes in the Premiere plug-in.

It turns out this was due to bugs in Xcode 12.2 and earlier.  Updating
to Xcode 12.5.1 fixed the problem and this commit was probably
unnecessary.