Fix alpha channel encoder bug
Was assuming that the RGB and Alpha encoders would spit out packets at
the same rate, which is not always the case. My solution is to hold
packets in queue objects, which is not ideal as it involves allocating
memory and copying buffers around, even when there is no alpha.